+++ src/Products/PloneMeeting/events.py
@@ -1,13 +1,4 @@
 # -*- coding: utf-8 -*-
-#
-# File: events.py
-#
-# Copyright (c) 2015 by Imio.be
-# Generator: ArchGenXML Version 2.7
-#            http://plone.org/products/archgenxml
-#
-# GNU General Public License (GPL)
-#
 
 from AccessControl import Unauthorized
 from collective.contact.plonegroup.utils import get_all_suffixes
@@ -242,6 +233,7 @@ def onOrgWillBeRemoved(current_org, event):
       - it can not be referenced in an existing MeetingConfig;
       - it can not be used in an existing MeetingCategory.usingGroups;
       - it can not be used as groupInCharge of another organization;
+      - it can not be used as groupInCharge of a category;
       - the linked ploneGroups must be empty of members.'''
     # Do lighter checks first...  Check that the organization is not used
     # in a meetingConfig
@@ -284,7 +276,7 @@ def onOrgWillBeRemoved(current_org, event):
         categories = mc.categories.objectValues('MeetingCategory')
         classifiers = mc.classifiers.objectValues('MeetingCategory')
         for cat in tuple(categories) + tuple(classifiers):
-            if current_org_uid in cat.getUsingGroups():
+            if current_org_uid in cat.getUsingGroups() or current_org_uid in cat.getGroupsInCharge():
                 raise BeforeDeleteException(translate("can_not_delete_organization_meetingcategory",
                                                       mapping={'url': cat.absolute_url()},
                                                       domain="plone",