+++ src/Products/PloneMeeting/indexes.py
@@ -158,6 +158,14 @@ def linkedMeetingDate(obj):
 
 
 @indexer(IMeetingItem)
+def getGroupsInCharge(obj):
+    """
+      Indexes the groupsInCharge attribute including auto groups
+    """
+    return obj.getGroupsInCharge(includeAuto=True)
+
+
+@indexer(IMeetingItem)
 def getPreferredMeetingDate(obj):
     """
       Store the preferredMeeting date.
@@ -204,10 +212,9 @@ def sentToInfos(obj):
     return res or _marker
 
 
-@indexer(IMeetingItem)
 def SearchableText(obj):
     """
-      Contained annex title is indexed in the item's SearchableText.
+      Contained annex title is indexed in the SearchableText.
     """
     res = []
     res.append(obj.SearchableText())
@@ -215,6 +222,16 @@ def SearchableText(obj):
         res.append(annex.SearchableText())
     res = ' '.join(res)
     return res or _marker
+
+
+@indexer(IMeetingItem)
+def SearchableText_item(obj):
+    return SearchableText(obj)
+
+
+@indexer(IMeeting)
+def SearchableText_meeting(obj):
+    return SearchableText(obj)
 
 
 @indexer(IMeetingItem)