+++ src/Products/PloneMeeting/vocabularies.py
@@ -268,8 +268,9 @@ class GroupsInChargeVocabulary(object):
         cfg = tool.getMeetingConfig(context)
         res = []
         is_using_cfg_order = False
-        # groups in charge are defined on organizations
         if 'groupsInCharge' not in cfg.getUsedItemAttributes():
+            # groups in charge are defined on organizations or categories
+            # organizations
             orgs = get_organizations(only_selected=only_selected)
             for org in orgs:
                 for group_in_charge_uid in (org.groups_in_charge or []):
@@ -277,6 +278,16 @@ class GroupsInChargeVocabulary(object):
                     # manage duplicates
                     if group_in_charge and group_in_charge not in res:
                         res.append(group_in_charge)
+            # categories
+            if not cfg.getUseGroupsAsCategories():
+                classifiers = 'classifier' in cfg.getUsedItemAttributes()
+                categories = cfg.getCategories(classifiers=classifiers, onlySelectable=False, caching=False)
+                for cat in categories:
+                    for group_in_charge_uid in cat.getGroupsInCharge():
+                        group_in_charge = get_organization(group_in_charge_uid)
+                        # manage duplicates
+                        if group_in_charge and group_in_charge not in res:
+                            res.append(group_in_charge)
         else:
             # groups in charge are selected on the items
             is_using_cfg_order = True
@@ -611,6 +622,8 @@ class AskedAdvicesVocabulary(object):
             self.cfg = self.tool.getMeetingConfig(context)
         except:
             return SimpleVocabulary(res)
+        if self.cfg is None:
+            return SimpleVocabulary(res)
 
         self.context = context
         self.request = context.REQUEST
@@ -830,7 +843,7 @@ class SentToInfosVocabulary(object):
                    )
         for cfgInfo in cfg.getMeetingConfigsToCloneTo():
             cfgId = cfgInfo['meeting_config']
-            cfgTitle = getattr(tool, cfgId).getName()
+            cfgTitle = getattr(tool, cfgId).Title()
             # add 'clonable to' and 'cloned to' options
             for suffix in ('__clonable_to', '__clonable_to_emergency',
                            '__cloned_to', '__cloned_to_emergency'):